free node cache for rbtree, fix transplant
The red-black tree is basically used as a queue, so adding a cache
to hold free nodes saves the allocator and GC a lot of work.

Doing this exposes a problem in transplant: it is sometimes called
with the t.none sentinel node as the source, which causes panics
now that nodes are cleared without regard for the sentinel node's
parent pointer. Previously this problem never occured since GC would
keep the parent around until it got set to something else. As it turns
out the old implementation's whole effort to update the source's parent
is unnecessary anyways, so it is gone now.

Change-Id: I7284d51b5c9ce17e730d259c6a3959c2f13b302f
Also: more tests!